# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: NextSmartShip openapi
  description: NextSmartShip openapi for oms and wms
  version: ""
host: 0.0.0.0:80
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /inbound-orders:
    get:
      tags:
      - order
      summary: get_inbound_order order
      operationId: order#get_inbound_order
      parameters:
      - name: client_order_id
        in: query
        description: client_order_id
        required: true
        type: string
        minLength: 1
      responses:
        "200":
          description: OK response.
          schema:
            type: array
            items:
              $ref: '#/definitions/InboundOrderResponseResponse'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/OrderGetInboundOrderUnauthorizedResponseBody'
      schemes:
      - http
    put:
      tags:
      - order
      summary: update_inbound_order order
      operationId: order#update_inbound_order
      parameters:
      - name: update_inbound_order_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/OrderUpdateInboundOrderRequestBody'
          required:
          - customer_order_id
          - warehouse_id
          - customer_code
          - tracking_number
          - requested_pickup_at
          - estimated_arrival_at
          - items
          - type
          - address
          - is_pickup
          - description
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/OrderUpdateInboundOrderResponseBody'
            required:
            - status
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/OrderUpdateInboundOrderUnauthorizedResponseBody'
      schemes:
      - http
    post:
      tags:
      - order
      summary: create_inbound_order order
      operationId: order#create_inbound_order
      parameters:
      - name: create_inbound_order_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/OrderCreateInboundOrderRequestBody'
          required:
          - customer_order_id
          - warehouse_id
          - customer_code
          - tracking_number
          - requested_pickup_at
          - estimated_arrival_at
          - items
          - type
          - address
          - is_pickup
          - description
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/OrderCreateInboundOrderResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/OrderCreateInboundOrderUnauthorizedResponseBody'
      schemes:
      - http
...
  /outbound-orders:
    get:
      tags:
      - order
      summary: get_outbound_order order
      operationId: order#get_outbound_order
      parameters:
      - name: client_order_id
        in: query
        description: client_order_id
        required: true
        type: string
        minLength: 1
      responses:
        "200":
          description: OK response.
          schema:
            type: array
            items:
              $ref: '#/definitions/OrderRspResponse'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/OrderGetOutboundOrderUnauthorizedResponseBody'
      schemes:
      - http
    put:
      tags:
      - order
      summary: update_outbound_order order
      operationId: order#update_outbound_order
      parameters:
      - name: update_outbound_order_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/OrderUpdateOutboundOrderRequestBody'
          required:
          - customer_order_id
          - customer_code
          - total_price
          - currency
          - customer_tariff_number
          - customer_tariff_number_type
          - enable_prepay_tariff
          - shipping_type
          - receiver_info
          - items
          - channel_id
          - description
          - inbound_order_id
          - type
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/OrderUpdateOutboundOrderResponseBody'
            required:
            - status
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/OrderUpdateOutboundOrderUnauthorizedResponseBody'
      schemes:
      - http
    post:
      tags:
      - order
      summary: create_outbound_order order
      operationId: order#create_outbound_order
      parameters:
      - name: create_outbound_order_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/OrderCreateOutboundOrderRequestBody'
          required:
          - customer_order_id
          - customer_code
          - total_price
          - currency
          - customer_tariff_number
          - customer_tariff_number_type
          - enable_prepay_tariff
          - shipping_type
          - receiver_info
          - items
          - channel_id
          - description
          - inbound_order_id
          - type
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/OrderCreateOutboundOrderResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/OrderCreateOutboundOrderUnauthorizedResponseBody'
      schemes:
      - http
  /pickup-orders:
    post:
      tags:
      - order
      summary: create_pickup_order order
      operationId: order#create_pickup_order
      parameters:
      - name: create_pickup_order_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/OrderCreatePickupOrderRequestBody'
          required:
          - requested_pickup_at
          - items
          - address
          - type
          - customer_code
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/OrderCreatePickupOrderResponseBody'
            required:
            - error_msg
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/OrderCreatePickupOrderUnauthorizedResponseBody'
      schemes:
      - http
...
definitions:
...
  InboundOrderResponseResponse:
    title: InboundOrderResponseResponse
    type: object
    properties:
      client_order_id:
        type: string
        description: client order id
        example: xxx1234
        maxLength: 50
      items:
        type: array
        items:
          $ref: '#/definitions/ItemResponse'
        description: items
        example:
        - barcode: YCrankshaft
          product_name: NSS Mate 40E
          product_sku: YCrankshaft
          qty: 3
        - barcode: YCrankshaft
          product_name: NSS Mate 40E
          product_sku: YCrankshaft
          qty: 3
        - barcode: YCrankshaft
          product_name: NSS Mate 40E
          product_sku: YCrankshaft
          qty: 3
      platform_order_id:
        type: integer
        description: platform order id
        example: 123
        format: int64
      status:
        type: integer
        description: order status(1 准备揽件 2 运输中 3 已到库)
        example: 1
        format: int64
      timestamp:
        type: string
        description: timestamp
        example: "2022-01-21 02:26:54"
      tracking_number:
        type: string
        description: tracking number
        example: Qui explicabo accusamus.
      tracking_url:
        type: string
        description: tracking url
        example: Inventore minus enim iure.
    description: OrderRsp describes the order info
    example:
      client_order_id: xxx1234
      items:
      - barcode: YCrankshaft
        product_name: NSS Mate 40E
        product_sku: YCrankshaft
        qty: 3
      - barcode: YCrankshaft
        product_name: NSS Mate 40E
        product_sku: YCrankshaft
        qty: 3
      platform_order_id: 123
      status: 1
      timestamp: "2022-01-21 02:26:54"
      tracking_number: Natus deserunt quidem in sint distinctio non.
      tracking_url: Occaecati consequatur iusto perspiciatis.
    required:
    - client_order_id
    - status
    - platform_order_id
    - tracking_number
    - tracking_url
    - items
    - timestamp
...
  OrderGetInboundOrderUnauthorizedResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: get_inbound_order_Unauthorized_response_body result type (default
      view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
...
  OrderUpdateInboundOrderRequestBody:
    title: OrderUpdateInboundOrderRequestBody
    type: object
    properties:
      address:
        $ref: '#/definitions/ShippingAddressRequestBody'
      customer_code:
        type: string
        description: customer code
        example: "1013"
        minLength: 1
        maxLength: 50
      customer_order_id:
        type: string
        description: Customer Order ID
        example: xxx1234
        minLength: 1
        maxLength: 50
      description:
        type: string
        description: description
        example: description
      estimated_arrival_at:
        type: string
        description: estimated arrival at
        example: "2021-01-12 09:34:09"
        maxLength: 30
      id:
        type: integer
        description: inbound order id
        example: 1
        format: int32
      is_pickup:
        type: boolean
        description: is pickup
        example: true
      items:
        type: array
        items:
          $ref: '#/definitions/ItemRequestBody'
        description: inbound order items
        example:
        - barcode: YCrankshaft
          product_name: NSS Mate 40E
          product_sku: YCrankshaft
          qty: 3
        - barcode: YCrankshaft
          product_name: NSS Mate 40E
          product_sku: YCrankshaft
          qty: 3
      requested_pickup_at:
        type: string
        description: requested pickup at
        example: "2021-01-12 09:34:09"
        maxLength: 30
      tracking_number:
        type: string
        description: tracking number
        example: YT000001
        maxLength: 50
      type:
        type: integer
        description: delivery mode(1 direct，2 warehouse)
        example: 1
        enum:
        - 1
        - 2
      warehouse_id:
        type: integer
        description: warehouse id
        example: 1
        minimum: 1
    example:
      address:
        address1: address1
        address2: address2
        city_name: SZ
        country_code: US
        country_name: US
        first_name: He
        last_name: John
        name: He
        phone_number: 150xxxxxxxx
        state_code: S
        state_name: U
        zip_code: "10016"
      customer_code: "1013"
      customer_order_id: xxx1234
      description: description
      estimated_arrival_at: "2021-01-12 09:34:09"
      id: 1
      is_pickup: true
      items:
      - barcode: YCrankshaft
        product_name: NSS Mate 40E
        product_sku: YCrankshaft
        qty: 3
      - barcode: YCrankshaft
        product_name: NSS Mate 40E
        product_sku: YCrankshaft
        qty: 3
      - barcode: YCrankshaft
        product_name: NSS Mate 40E
        product_sku: YCrankshaft
        qty: 3
      - barcode: YCrankshaft
        product_name: NSS Mate 40E
        product_sku: YCrankshaft
        qty: 3
      requested_pickup_at: "2021-01-12 09:34:09"
      tracking_number: YT000001
      type: 1
      warehouse_id: 1
    required:
    - customer_order_id
    - warehouse_id
    - customer_code
    - tracking_number
    - requested_pickup_at
    - estimated_arrival_at
    - items
    - type
    - address
    - is_pickup
    - description
  OrderUpdateInboundOrderResponseBody:
    title: OrderUpdateInboundOrderResponseBody
    type: object
    properties:
      status:
        type: integer
        description: status
        example: 1
        enum:
        - 0
        - 1
    example:
      status: 1
    required:
    - status
  OrderUpdateInboundOrderUnauthorizedResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: update_inbound_order_Unauthorized_response_body result type (default
      view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
...